+++ smart-pos-core/src/main/java/com/interswitchng/smartpos/modules/card/CardViewModel.kt
@@ -48,7 +48,7 @@ internal class CardViewModel(private val posDevice: POSDevice, private val isoSe
 
     fun getCardPAN() = emv.getPan()
 
-    fun setupTransaction(amount: Int, terminalInfo: TerminalInfo) {
+    fun setupTransaction(amount: Double, terminalInfo: TerminalInfo) {
 
         with(uiScope) {
             // launch job in IO thread to listen for messages
