+++ smart-pos-core/src/main/java/com/interswitchng/smartpos/shared/models/transaction/PaymentInfo.kt
@@ -11,14 +11,14 @@ import com.interswitchng.smartpos.IswPos
  */
 class PaymentInfo: Parcelable {
 
-    val amount: Int
+    val amount: Double
     var bankCode: String
     lateinit var currentStan: String
 
     var originalStanId:String? = ""
     var originalAuthId:String? = ""
 
-    constructor(amount: Int, bankCode: String?,originalStanId:String?="",originalAuthId:String?="") {
+    constructor(amount: Double, bankCode: String?,originalStanId:String?="",originalAuthId:String?="") {
         this.amount = amount
         this.bankCode = bankCode  ?: ""
         this.originalStanId=originalStanId
@@ -26,7 +26,7 @@ class PaymentInfo: Parcelable {
     }
 
     constructor(parcel: Parcel) : this(
-            parcel.readInt(),
+            parcel.readDouble(),
             parcel.readString(),
             parcel.readString(),
             parcel.readString())
@@ -34,7 +34,7 @@ class PaymentInfo: Parcelable {
     fun getStan() = IswPos.getNextStan().also { currentStan = it }
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
-        parcel.writeInt(amount)
+        parcel.writeDouble(amount)
         parcel.writeString(bankCode)
         parcel.writeString(originalStanId)
         parcel.writeString(originalAuthId)
